requirements.txt
gradio
#google-cloud-aiplatform
langchain_google_vertexai
#langchain
python-dotenv